auto-generate template.json
