+++ .github/workflows/ci.yml
@@ -18,4 +18,4 @@ jobs:
         run: npm test
 
       - name: Compile
-        run: NODE_ENV=production npx rollup -c
+        run: make
