+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 dist
 node_modules
+template.json
