+++ src/data/config.ts
@@ -1,0 +1,14 @@
+import { CharacterSourceData } from "./CharacterSourceData"
+
+declare global
+{
+    interface SourceConfig
+    {
+        Actor: { type: 'character', data: CharacterSourceData }
+    }
+
+    interface DataConfig
+    {
+        Actor: { type: 'character', data: CharacterSourceData }
+    }
+}
