--- template.json
@@ -1,28 +1,0 @@
-{
-    "Actor": {
-        "types": [
-            "character"
-        ],
-        "character": {
-            "attributes": {
-                "hp": {
-                    "value": 0,
-                    "max": 0
-                },
-                "init": 0,
-                "size": {
-                    "category": "medium",
-                    "reach": 5
-                },
-                "speed": {
-                    "base": 30
-                }
-            },
-            "heroLabSync": {
-                "lastUpdate": 0,
-                "file": "",
-                "character": ""
-            }
-        }
-    }
-}
